Extract beep function
