+++ frontend/src/index.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
+import beep from './beep';
 import Errors from './Errors';
 
 const e = React.createElement;
@@ -1705,23 +1706,6 @@ ws.onmessage = (event) => {
 
   renderUI();
 };
-
-let beepContext: AudioContext | null = null;
-
-function beep(vol: number, freq: number, duration: number) {
-  if (!beepContext) {
-    beepContext = new AudioContext();
-  }
-  const v = beepContext.createOscillator();
-  const u = beepContext.createGain();
-  v.connect(u);
-  v.frequency.value = freq;
-  v.type = 'square';
-  u.connect(beepContext.destination);
-  u.gain.value = vol * 0.01;
-  v.start(beepContext.currentTime);
-  v.stop(beepContext.currentTime + duration * 0.001);
-}
 
 declare var CARDS: ICardInfo[];
